# openapi.yaml
openapi: 3.1.0
info:
  title: Traccar
  description: Traccar GPS tracking server API documentation. To use the API you need to have a server instance. For testing purposes you can use one of free [demo servers](https://www.traccar.org/demo-server/). For production use you can install your own server or get a [subscription service](https://www.traccar.org/product/tracking-server/).
  contact:
    name: Traccar Support
    url: https://www.traccar.org/
    email: support@traccar.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '6.7.5-GaC'
servers:
  - url: https://demo.traccar.org/api
    description: Demo Server 1
  - url: https://demo2.traccar.org/api
    description: Demo Server 2
  - url: https://demo3.traccar.org/api
    description: Demo Server 3
  - url: https://demo4.traccar.org/api
    description: Demo Server 4
  - url: https://server.traccar.org/api
    description: Subscription Server
  - url: http://{host}:{port}/api
    description: Other Server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '8082'
          - '80'
        default: '8082'
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
...
  /geofences:
    get:
      summary: Fetch a list of Geofences
      description: Without params, it returns a list of Geofences the user has access to
      tags:
        - Geofences
      parameters:
        - name: all
          in: query
          description: Can only be used by admins or managers to fetch all entities
          schema:
            type: boolean
        - name: userId
          in: query
          description: Standard users can use this only with their own _userId_
          schema:
            type: integer
        - name: deviceId
          in: query
          description: >-
            Standard users can use this only with _deviceId_s, they have access
            to
          schema:
            type: integer
        - name: groupId
          in: query
          description: >-
            Standard users can use this only with _groupId_s, they have access
            to
          schema:
            type: integer
        - name: refresh
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Geofence'
    post:
      summary: Create a Geofence
      tags:
        - Geofences
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Geofence'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Geofence'
      x-codegen-request-body-name: body
...
  /reports/route:
    get:
      summary: >-
        Fetch a list of Positions within the time period for the Devices or
        Groups
      description: At least one _deviceId_ or one _groupId_ must be passed
      tags:
        - Reports
      parameters:
        - name: deviceId
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: integer
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Position'
  /reports/events:
    get:
      summary: Fetch a list of Events within the time period for the Devices or Groups
      description: At least one _deviceId_ or one _groupId_ must be passed
      tags:
        - Reports
      parameters:
        - name: deviceId
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: integer
        - name: type
          in: query
          description: '% can be used to return events of all types'
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
  /reports/geofences:
    get:
      summary: >-
        Fetch geofence enter/exit intervals within the time period for Devices
        or Groups
      description: At least one _deviceId_ or one _groupId_ must be passed
      tags:
        - Reports
      parameters:
        - name: deviceId
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: integer
        - name: geofenceId
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: integer
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportGeofences'
  /reports/summary:
    get:
      summary: >-
        Fetch a list of ReportSummary within the time period for the Devices or
        Groups
      description: At least one _deviceId_ or one _groupId_ must be passed
      tags:
        - Reports
      parameters:
        - name: deviceId
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: integer
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportSummary'
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportSummary'
  /reports/trips:
    get:
      summary: >-
        Fetch a list of ReportTrips within the time period for the Devices or
        Groups
      description: At least one _deviceId_ or one _groupId_ must be passed
      tags:
        - Reports
      parameters:
        - name: deviceId
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: integer
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportTrips'
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportTrips'
  /reports/stops:
    get:
      summary: >-
        Fetch a list of ReportStops within the time period for the Devices or
        Groups
      description: At least one _deviceId_ or one _groupId_ must be passed
      tags:
        - Reports
      parameters:
        - name: deviceId
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: integer
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportStops'
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportStops'
...
components:
  schemas:
    Position:
      type: object
      properties:
        id:
          type: integer
          format: int64
        deviceId:
          type: integer
          format: int64
        protocol:
          type: string
        deviceTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        fixTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        serverTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        outdated:
          type: boolean
        valid:
          type: boolean
        latitude:
          type: number
        longitude:
          type: number
        altitude:
          type: number
        speed:
          type: number
          description: in knots
        course:
          type: number
        address:
          type: string
        accuracy:
          type: number
        network:
          type: object
          properties: {}
        geofenceIds:
          type: array
          items:
            type: integer
        attributes:
          type: object
          properties: {}
...
    Geofence:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        area:
          type: string
        calendarId:
          type: integer
          format: int64
        attributes:
          type: object
          properties: {}
...
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
        eventTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        deviceId:
          type: integer
          format: int64
        positionId:
          type: integer
          format: int64
        geofenceId:
          type: integer
          format: int64
        maintenanceId:
          type: integer
        attributes:
          type: object
          properties: {}
    ReportSummary:
      type: object
      properties:
        deviceId:
          type: integer
          format: int64
        deviceName:
          type: string
        maxSpeed:
          type: number
          description: in knots
        averageSpeed:
          type: number
          description: in knots
        distance:
          type: number
          description: in meters
        spentFuel:
          type: number
          description: in liters
        engineHours:
          type: integer
    ReportTrips:
      type: object
      properties:
        deviceId:
          type: integer
          format: int64
        deviceName:
          type: string
        maxSpeed:
          type: number
          description: in knots
        averageSpeed:
          type: number
          description: in knots
        distance:
          type: number
          description: in meters
        spentFuel:
          type: number
          description: in liters
        duration:
          type: integer
        startTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        startAddress:
          type: string
        startLat:
          type: number
        startLon:
          type: number
        endTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        endAddress:
          type: string
        endLat:
          type: number
        endLon:
          type: number
        driverUniqueId:
          type: string
        driverName:
          type: string
    ReportStops:
      type: object
      properties:
        deviceId:
          type: integer
          format: int64
        deviceName:
          type: string
        duration:
          type: integer
        startTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        address:
          type: string
        lat:
          type: number
        lon:
          type: number
        endTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        spentFuel:
          type: number
          description: in liters
        engineHours:
          type: integer
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer